day4/pyproject.toml
[project]
name = "day4"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "python-dotenv",
    "google-cloud-aiplatform",
]
